Project_Ergun_Yilmaz/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
ca-certificates=2022.6.15=h5b45459_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
certifi=2021.5.30=py36ha15d459_0
cloudpickle=2.1.0=pyhd8ed1ab_0
colorama=0.4.5=pyhd8ed1ab_0
cudatoolkit=11.0.3=h3f58a73_10
cycler=0.11.0=pyhd8ed1ab_0
cytoolz=0.11.0=py36h68aa20f_3
dask-core=2021.3.0=pyhd8ed1ab_0
dataclasses=0.8=pyh787bdff_2
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
freetype=2.10.4=h546665d_1
h5py=3.1.0=nompi_py36hf359dfe_100
hdf5=1.10.6=nompi_h5268f04_1114
icc_rt=2019.0.0=h0cc432a_1
icu=69.1=h0e60522_0
imagecodecs-lite=2019.12.3=py36h6434af4_3
imageio=2.13.1=pyhd8ed1ab_0
intel-openmp=2022.1.0=h57928b3_3787
ipykernel=5.5.5=py36hfacbf0b_0
ipython=7.16.1=py36h7b2dad6_2
ipython_genutils=0.2.0=py_1
jedi=0.17.2=py36ha15d459_1
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=h8ffe710_2
jupyter_client=7.1.2=pyhd8ed1ab_0
jupyter_core=4.8.1=py36ha15d459_0
kiwisolver=1.3.1=py36he95197e_1
krb5=1.19.3=h1176d77_0
lerc=3.0=h0e60522_0
libblas=3.9.0=8_mkl
libcblas=3.9.0=8_mkl
libclang=13.0.1=default_h81446c8_0
libcurl=7.83.1=h789b8ee_0
libdeflate=1.10=h8ffe710_0
liblapack=3.9.0=8_mkl
libpng=1.6.37=h1d00b33_2
libsodium=1.0.18=h8d14728_1
libssh2=1.10.0=h680486a_2
libtiff=4.3.0=hc4061b1_4
libuv=1.43.0=h8ffe710_0
libzlib=1.2.12=h8ffe710_1
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py36hfa737b6_1
mkl=2020.4=hb70f87d_311
mkl-service=2.3.0=py36h779f372_2
mkl_fft=1.3.0=py36h6434af4_1
mkl_random=1.2.0=py36hd7f5668_1
msys2-conda-epoch=20160418=1
munch=2.5.0=pypi_0
nest-asyncio=1.5.5=pyhd8ed1ab_0
networkx=2.5=py_0
ninja=1.11.0=h2d74725_0
numpy=1.19.5=py36h4b40d73_2
numpy-base=1.18.5=py36hc3f5095_0
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1p=h8ffe710_0
pandas=1.1.5=py36he38d939_0
parso=0.7.1=pyh9f0ad1d_0
pickleshare=0.7.5=py_1003
pillow=5.2.0=py36h9a613e6_1
pip=21.3.1=pyhd8ed1ab_0
pretrainedmodels=0.7.4=pypi_0
prompt-toolkit=3.0.30=pyha770c72_0
pygments=2.12.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.12.3=py36ha15d459_7
pyqt-impl=5.12.3=py36he2d232f_7
pyqt5-sip=4.19.18=py36he2d232f_7
pyqtchart=5.12=py36he2d232f_7
pyqtwebengine=5.12.1=py36he2d232f_7
pyreadline=2.1=py36ha15d459_1004
python=3.6.15=h39d44d4_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.6=2_cp36m
pytorch=1.7.0=py3.6_cuda110_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pyhd8ed1ab_0
pywavelets=1.1.1=py36h14b15b0_3
pywin32=301=py36h68aa20f_0
pyyaml=5.4.1=py36h68aa20f_1
pyzmq=22.3.0=py36h1d5d788_0
qt=5.12.9=h556501e_6
scikit-image=0.17.2=py36he38d939_4
scikit-learn=0.24.2=py36h5a2dbc3_1
scipy=1.5.3=py36h27d303f_1
setuptools=58.0.4=py36ha15d459_2
sip=4.19.25=py36he2d232f_0
six=1.16.0=pyh6c4a22f_0
spyder-kernels=2.1.0=py36ha15d459_0
sqlite=3.39.0=h8ffe710_0
threadpoolctl=3.1.0=pyh8a188c0_0
tifffile=2020.6.3=py_0
tk=8.6.12=h8ffe710_0
toolz=0.11.2=pyhd8ed1ab_0
torchvision=0.8.1=py36_cu110
tornado=6.1=py36h68aa20f_1
tqdm=4.26.0=py_0
traitlets=4.3.3=pyhd8ed1ab_2
typing_extensions=4.1.1=pyha770c72_0
ucrt=10.0.20348.0=h57928b3_0
vc=14.2=hb210afc_6
vision=1.0.0=pypi_0
vs2015_runtime=14.29.30037=h902a5da_6
wcwidth=0.2.5=pyh9f0ad1d_2
wheel=0.37.1=pyhd8ed1ab_0
wincertstore=0.2=py36ha15d459_1006
xz=5.2.5=h62dcd97_1
yaml=0.2.5=h8ffe710_2
zeromq=4.3.4=h0e60522_1
zlib=1.2.12=h8ffe710_1
zstd=1.5.2=h6255e5f_2
